requirements.txt
flask
requests
gunicorn
psycopg2-binary
sqlalchemy
openai
google-generativeai